+++ src/shared/ui/card.tsx
@@ -1,7 +1,7 @@
 
 import * as React from "react"
 
-import { cn } from "@/shared/utils"
+import { cn } from "@/lib/utils"
 
 const Card = React.forwardRef<
   HTMLDivElement,
