+++ src/shared/ui/input.tsx
@@ -1,7 +1,7 @@
 
 import * as React from "react"
 
-import { cn } from "@/shared/utils"
+import { cn } from "@/lib/utils"
 
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {}
